pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.poetry]
name = "research-experiment"
version = "0.1.0"
description = "Hydra-driven research experiments"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
torch = "*"
torchvision = "*"
hydra-core = "^1.3"
optuna = "*"
transformers = "*"
wandb = "*"
pandas = "*"
matplotlib = "*"

[tool.poetry.scripts]
run-experiment = "src.main:main"
